move the zuora models out of the soap namespace
