+++ membership-common/src/main/scala/com/gu/zuora/models/PaymentSummary.scala
@@ -1,9 +1,9 @@
-package com.gu.zuora.soap.models
+package com.gu.zuora.models
 
 import com.gu.i18n.Currency
 import com.gu.memsub.Price
 import com.gu.stripe.Stripe
-import com.gu.zuora.soap.models.Queries.{InvoiceItem, PreviewInvoiceItem, Subscription}
+import com.gu.zuora.models.Queries.{InvoiceItem, PreviewInvoiceItem, Subscription}
 
 case class PaymentSummary(current: InvoiceItem, previous: Seq[InvoiceItem], currency: Currency) {
   val totalPrice = current.price + previous.map(_.price).sum
