+++ membership-common/src/test/scala/com/gu/zuora/ZuoraServiceTest.scala
@@ -1,7 +1,7 @@
 package com.gu.zuora
 
 import com.github.nscala_time.time.Imports._
-import com.gu.zuora.soap.models.Queries.{Amendment, InvoiceItem, Subscription}
+import com.gu.zuora.models.Queries.{Amendment, InvoiceItem, Subscription}
 import org.joda.time.{DateTime, DurationFieldType}
 import org.specs2.mutable.Specification
